app/requirements.txt
dash
plotly
pandas
flask
gunicorn
google-cloud-storage
dash-bootstrap-components
psutil
google-cloud-bigquery
google-cloud-bigquery-storage
db-dtypes